vscode-cue-visualizer/bundled/pyproject.toml
[tool.poetry]
name = "blarify"
version = "1.2.0"
description = "A simple graph builder based on LSP calls"
authors = ["Juan Vargas <juan@blar.io>", "Benjamín Errazuriz <benjamin@blar.io>"]
readme = "README.md"
homepage = "https://blar.io"
repository = "https://github.com/blarApp/blarify"
license = "MIT"
packages = [{include = "blarify"}]

[tool.poetry.dependencies]
python = ">=3.10,<=3.14"
asyncio = "^3.4.3"
websockets = "^13.1"
neo4j = "^5.25.0"
python-dotenv = "^1.0.1"
tree-sitter = "^0.23.2"
tree-sitter-python = "^0.23.2"
tree-sitter-ruby = "^0.23.0"
tree-sitter-javascript = "^0.23.0"
tree-sitter-typescript = "^0.23.2"
tree-sitter-c-sharp = "^0.23.1"
tree-sitter-go = "^0.23.4"
falkordb = "^1.0.10"
requests = "^2.32.3"
typing-extensions = "^4.12.2"
jedi-language-server = "^0.43.1"
psutil = "^7.0.0"
vendoring = "^1.2.0"
tree-sitter-php = "^0.23.11"
tree-sitter-java = "^0.23.5"
openai = "^1.35.0"
pathspec = "^0.12.1"



[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.1.0"
pyright = "^1.1.403"

[tool.vendoring]
destination = "blarify/vendor"
namespace = "blarify.vendor"
requirements = "requirements-vendor.txt"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["vendor/"]

# Format configuration
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Linting configuration (includes import sorting via "I" rules)
[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "COM812", "ISC001"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["S101", "PLR2004"]

[tool.pyright]
include = ["blarify/**/*.py"]
exclude = ["vendor/**", "build/**", ".venv/**"]
typeCheckingMode = "strict"
pythonVersion = "3.10"
pythonPlatform = "All"
reportMissingImports = true
reportMissingTypeStubs = false
useLibraryCodeForTypes = true

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

# Note: isort is no longer needed as Ruff handles import sorting
[tool.isort]
profile = "black"
